LPS-77447 Move com.liferay.mobile.device.rules.rule.group.action package to web module
